// frontend/lib/schemas.ts
import { z } from 'zod';

export const createUserSchema = z.object({
    firstName: z.string().min(1, 'First name is required'),
    lastName: z.string().min(1, 'Last name is required'),
    email: z.email('Invalid email address'),
    jobTitle: z.string().optional(),
    role: z.enum(['Admin', 'Supervisor', 'Operator', 'Client']),
    supervisor: z.string().optional(),
    password: z.string().min(8, 'Password must be at least 8 characters'),
    status: z.boolean().optional(),
});

export type CreateUserFormData = z.infer<typeof createUserSchema>;

export const editUserSchema = z.object({
    firstName: z.string().min(1, 'First name is required'),
    lastName: z.string().min(1, 'Last name is required'),
    email: z.string().email('Invalid email address'),
    role: z.enum(['Admin', 'Supervisor', 'Operator', 'Client']),
    supervisor: z.string().optional(),
    status: z.boolean().optional(),
});

export type EditUserFormData = z.infer<typeof editUserSchema>;